Enable Hot Module Reloading for quicker change refresh and better dev-experience (#98)
Currently, any file within the /static/client/* will trigger build-js and build-css. This collectively takes around 8-10 seconds to show up the change in the browser, and that too after a page refresh. For developers, it is not a good/productive experience and the wait is annoying.

The solution is to enable hot module reloading(HMR) by Vite, and since Vite is already being used for builds, we only need to enable HMR for dev environment.

Co-authored-by: Muhammad Ali <muhammad.ali@canonical.com>